Add getFeeStats method

+++ cmd/soroban-rpc/internal/methods/get_events.go
@@ -295,7 +295,7 @@ type PaginationOptions struct {
 
 type GetEventsResponse struct {
 	Events       []EventInfo `json:"events"`
-	LatestLedger int64       `json:"latestLedger"`
+	LatestLedger uint32      `json:"latestLedger"`
 }
 
 type eventScanner interface {
@@ -372,7 +372,7 @@ func (h eventsRPCHandler) getEvents(request GetEventsRequest) (GetEventsResponse
 		results = append(results, info)
 	}
 	return GetEventsResponse{
-		LatestLedger: int64(latestLedger),
+		LatestLedger: uint32(latestLedger),
 		Events:       results,
 	}, nil
 }
